requirements.txt
PyQt6
biopython
qdarkstyle
